big changes to table writing, covariates in soft-clipped reads, and skip finding
 * insertions will be skipped if they appear between skipped sites
 * EmpiricalQuality based on EstimatedQReported, not meanq
 * 0 data removed from RG table
 * 0 data removed from dinuc table
 * cycle covariate ignores soft clipped bases
 * dinuc covariate ignores soft clipped bases

The cycle/dinuc covariate changes may need to be reverted at some point.
GATK hard-clips the reads before calculating covariates for building the
model, so for creating the table we need to ignore those bases.  However,
we had recalibration tests pass when we included them so gatk's
ApplyRecalibration may include them. This would require us to use 2 different
functions depending on whether we are building the model or applying the
recalibration.